Move SingleDatePicker to SingleDatePickerContainer
I am working to split up this component into two pieces so that folks
can use code splitting to reduce bundle sizes. To keep the diffs
readable, I am starting with this commit that moves most of the
SingleDatePicker into a new component that is rendered by
SingleDatePicker. In a future commit, I will split out more of it and
pass it in to the SingleDatePickerContainer.
